refactor: add `termination_callbacks` context
remove supervisors context and add termination_callbacks.

the old `TerminationCallbacks` is now a `Worker` and is only supposed
to be called by the supervisor.
the supervisor has been moved top level exposes the worker's `add` function
